ushort to byte transport modes

+++ Algorithms/ArrayDijkstra/ArrayDijkstra.cs
@@ -26,7 +26,7 @@ namespace SytyRouting.Algorithms.ArrayDijkstra
             }
         }
 
-        protected override List<Node> RouteSearch(Node originNode, Node destinationNode, ushort transportMode)
+        protected override List<Node> RouteSearch(Node originNode, Node destinationNode, byte transportMode)
         {
             route.Clear();
             routeCost = 0;
