+++ Algorithms/Dijkstra/Dijkstra.cs
@@ -19,7 +19,7 @@ namespace SytyRouting.Algorithms.Dijkstra
             }
         }
 
-        protected override List<Node> RouteSearch(Node originNode, Node destinationNode, ushort transportMode)
+        protected override List<Node> RouteSearch(Node originNode, Node destinationNode, byte transportMode)
         {
             route.Clear();
             routeCost = 0;
